01_etl/app/requirements.txt
bandit==1.7.2
django_extensions
django==3.2
django-debug-toolbar
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-split-settings==1.1
faker==13.3.4
flake8==4.0
flake8-bandit==2.1.2
gunicorn==20.0.4
pre-commit==2.15.0
psycopg2-binary==2.9
pytest==6.2.5
python-dotenv==0.20
wemake-python-styleguide==0.16.1
